Add Makefile task `github-preview`
This task will add 2 hidden files configuring the GitHub repository so that
a PDF preview is generated when a commit is pushed and when a PullRequest is
updated with commits. In the first case, the PDF preview will be published
in a GitHub Pre-Release (and will have a fix URL). In the second case, the PDF
will be accessible as a check artefact through the `Checks` tab of the
PullRequest.

The task `github-preview` will create two YAML files inside the directory
`.github/workflows/`:

  - `build.yml`   - workflow triggered at each PullRequest
  - `preview.yml` - workflow triggered at each pushed commit

Both files have to be commited and push to a remote GitHub repository to enable
the corresponding GitHub workflows.