Control the width of the main document and the architecture diagram in html.
The most intrusive change here is that the HTML document body width
will now be restricted to 21cm -- that's already too much for my taste, but
it keeps things remotely readable when people have their browser fullscreen.

Also, this commit recovers the style of the architecture diagram.
This is based on an xslt hack right now rather than trying
to declare a css class in the LaTeX, because there's no facility
for the latter at this point.  It's something we might want to have for other
figures one of these days, though.