More gracefully deal with submodules
This commit does two things:

(a) it changes the makefile template so that the ivoatex submodule
(and potentially any others) will automatically be initialised
if people forgot to do the recurse-submodules (which ought to address
issue #27).

(b) it adds an "update" target to the ivoatex makefile so people can
say "make update" now and then to fetch the latest ivoatex.