Adding a ucd macro.
Apart from a bit of uniform markup, this in particular enables breaking
UCDs at semicolons and dots.  As a side effect, atoms can be hyphenated,
too.

If this is merged, the corresponding definitions in EPN-TAP should be
removed; UCDList's ucd macro is different (at least for now) and probably
should stay that way because of the special requirements in UCD presentation
there.