Replacing Tabs with blanks in various files.
I have an odious habit of having Tabs of width 2 blanks almost
everywhere.  This usually doesn't work well in files that are edited
by other people.  Hence, this commit replaces tabs with 2 blanks in
XML files and with four blanks (PEP 8) in Python files.

This huge commit makes me wish we had at least a regression test in place.
Anyone up for writing a document that exercises as much of this as we can?