Make the double quote non-active in tth.
tth has support for the traditional German TeX control sequences opening with
a double quote.  Since plain LaTeX puts out some quote-like thing instead,
this is constant trouble when people fumble a double quote (instead of
`` and '') into LaTeX source.

This commit contains a patch against the tth source and the resulting tth.c.
Sorry the tth.c diff is so large -- that's how lex works.  If you really
want to inspect it, follow the instructions in UPDATING.