Expunging @webpage from ivoabib.
We have not defined that in our bst, and I'm not so wild about citing
web pages as to think about a good way to do that.  If you have to cite
web sites, I suppose writing that into howpublished is as good as about
anything else short of minting DOIs for them.

Also, dropping a record that's really too incomplete, and fixing a few
minor typos in ivoabib.