Fix some whitespace issues in macros
Macros whose definitions include newlines for cosmetic reasons can
in some cases result in unwanted significant whitespace in HTML output.
This addresses issue #115 as advised by @msdemlei;
for \specialterm, \todo and \ivoatex by allowing long lines in the
LaTeX source to avoid the line break, and for \auxiliaryurl
by stripping cosmetic whitespace at the start of a line
after a "%" character at the end of the previous line
(since otherwise in that case the source line would get very long).

I've tested that this fix has the desired effect for \specialterm
(via \xmlel in DALI), but not for the others.