Fix DOI link text.
This was sort-of correct before, _but_ failed to take account of the
fact that the `\url{...}` macro (from the `url` package) is overridden
by the `hyperref` package, to create a link.  This meant that the
generated DOI link included a link within the link, which didn't work
in PDF, and confused the hell out of tth.  Replaced with `hyperref`'s
`\nolinkurl` macro, which does the same escaping, but without creating
a link.