Trying to guess and format URLs in schmemadoc-generated source.
This is not so much to make them clickable; the problem is that these URIs
will not be broken by LaTeX if not marked up as URIs.  With the markup
(and the breaklink option that ivoatex has), there's much fewer URIs sticking
out of generated text -- and that's relevant because these overfull boxes
cannot be manually fixed.