check git is available in gitmeta.tex target
Add a line checking that the git command is present, i.e. the
make will fail if it's not.  Prevents silent production of an
uninformative gitmeta.tex file in absence of git.

Absence of git may seem like an unlikely eventuality for an
ivoatex build, but it has happened when building documents
from docker.