On make tag, only push the newly-made tag now.
This is because the preview workflow sets and moves a tag auto-pdf-preview
that will keep conflicting when people have longer-lived checkouts.

The way this is written what can happen is that for some reason the
server push of the tag has failed; if a user runs make tag again then,
the git tag will fail ("already exists", even if it's on the same commit)
and hence the push won't happen.  I've considered fixing this with an
-f on tag, but then I felt we should not automatically reset these tags.
So, it'll error out and people have to manually figure out what to do.