Fix to make tth compile work for gcc 15.
GCC 15 interprets the external and static declarations that tth has as
prototypes and then fails compilation without forcing it back to gnu17.
It would of course be preferable to fix the tth source; Ian is disinclined
to do so.  Perhaps we'll need to fork it one day.

For now, I'm going for the forcing.  This is a bit unfortunate in that we
are now chained to gcc unless people define CC and CFLAGS themselves.  I forsee
trouble on OSX...